requirements.txt
# ChroLens_Sentinel 依賴套件
# 自動生成於 2026-01-19

Pillow
packaging
pystray
ttkbootstrap
update_dialog
update_manager
webbrowser
